Moved Matadata to api folder (because it may be used by the game programmer)
