+++ src/crepe/CMakeLists.txt
@@ -4,7 +4,6 @@ target_sources(crepe PUBLIC
 	ComponentManager.cpp
 	Component.cpp
 	Collider.cpp
-	Metadata.cpp
 )
 
 target_sources(crepe PUBLIC FILE_SET HEADERS FILES
@@ -13,7 +12,6 @@ target_sources(crepe PUBLIC FILE_SET HEADERS FILES
 	ComponentManager.hpp
 	Component.h
 	Collider.h
-	Metadata.h
 )
 
 add_subdirectory(api)
