+++ src/crepe/api/CMakeLists.txt
@@ -11,6 +11,7 @@ target_sources(crepe PUBLIC
 	Texture.cpp
 	AssetManager.cpp
 	Sprite.cpp
+	Metadata.cpp
 )
 
 target_sources(crepe PUBLIC FILE_SET HEADERS FILES
@@ -28,4 +29,5 @@ target_sources(crepe PUBLIC FILE_SET HEADERS FILES
 	Texture.h 
 	AssetManager.h 
 	AssetManager.hpp
+	Metadata.h
 )
