+++ src/example/ecs.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
 #include "../crepe/ComponentManager.h"
-#include "../crepe/Metadata.h"
 #include "../crepe/api/GameObject.h"
+#include "../crepe/api/Metadata.h"
 #include "../crepe/api/Transform.h"
 
 using namespace crepe;
