+++ src/makefile
@@ -24,8 +24,8 @@ MAX += crepe/Component.h
 MAX += crepe/ComponentManager.cpp
 MAX += crepe/ComponentManager.h
 MAX += crepe/ComponentManager.hpp
-MAX += crepe/Metadata.cpp
-MAX += crepe/Metadata.h
+MAX += crepe/api/Metadata.cpp
+MAX += crepe/api/Metadata.h
 TODO += crepe/Particle.cpp
 TODO += crepe/Particle.h
 TODO += crepe/Position.h
